[HT-5] Fixed database, split tests
Теперь тесты работают с локальной базой данных, находящейся в `docker`-контейнере, они разделены на группы и исправлены так, чтобы они могли работать независимо.
