+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
   "extension": ["ts"],
-  "spec": "server/test/test.ts",
+  "spec": "server/test/test*.ts",
   "require": "ts-node/register"
 }
