+++ server/test/README.md
@@ -4,19 +4,19 @@
 ```sh
 yarn test
 ```
-Первые несколько секунд будет запускаться само приложение, потом начнут выполняться тесты. Они разделены на 3 основные группы: работа с партнёром, пользователем и хобби. Для каждого теста выводится результат его выполнения: если тест выполнен успешно, выводится зелёная галочка, его название и, если он выполнялся довольно долго, время выполнения; если тест не пройден — название и текст ошибки. Тесты могут выполняться долго!
+Первые несколько секунд будет разворачиваться база данных и запускаться само приложение, потом начнут выполняться тесты. Они разделены на 3 основные группы: работа с партнёром, пользователем и хобби. Разные группы тестов выполняются независимо друг от друга. Для каждого теста выводится результат его выполнения: если тест выполнен успешно, выводится зелёная галочка, его название и, если он выполнялся довольно долго, время выполнения; если тест не пройден — название и текст ошибки.
 
 В тестовом окружении выводятся только логи выполнения тестов, остальные отключены.
 
 ## Описание
 Все основные вещи, касающиеся тестов, сложены в директории `test`:
-- В файле `test.ts` описаны сами тесты;
+- Сами тесты описаны в файлах, начинающихся со слова "test". Сейчас их 3: `test_provider.ts`, `test_user.ts` и `test_hobby.ts`, в каждом из них описана соответствующая группа тестов.
 - Файл `sample_data.ts` хранит маленький набор тестовых данных, использующихся для проверки работы контроллеров: 2 партнёра, 2 пользователя, 6 хобби и данные для обновления записей.
-- В файле `utils.ts` описаны полезные функции, не относящиеся напрямую к тестированию, но позволяющие сделать код описания тестов более кратким и лаконичным.
+- В файле `utils.ts` описаны полезные функции, не относящиеся напрямую к тестированию, но позволяющие сделать код описания тестов более лаконичным.
 
-Для тестов используется отдельная база данных, данные для подключения к которой описаны в конфиге `test.yaml`. Файл `.mocharc.json` описывает опции для запуска тестового фреймворка `mocha`, использующегося вместе с `assertion`-библотекой `Chai` и её плагином `chai-http`, позволяющим просто работать с http-запросами.
+Для тестов используется отдельная база данных, разворачиваемая локально в `docker`-контейнере с помощью скрипта `start-db`. Файл `.mocharc.json` описывает опции для запуска тестового фреймворка `mocha`, использующегося вместе с `assertion`-библотекой `Chai` и её плагином `chai-http`, позволяющим просто работать с http-запросами.
 
-Для всех тестов выставлен таймаут в 15 секунд. Если он вдруг окажется слишком мал для какого-то конкретного теста/группы тестов, его можно отредактировать, прописав таймаут в начале теста/группы:
+Для всех тестов выставлен таймаут в 0,5 c. Если он вдруг окажется слишком мал для какого-то конкретного теста/группы тестов, его можно отредактировать, прописав таймаут в начале теста/группы:
 ```javascript
 this.timeout(нужный таймаут в мс);
 ```
