+++ tasks/start-watch
@@ -2,14 +2,8 @@
 
 CONTAINER=hobbytaste_mongodb
 
-echo "Setting up DB..."
 trap './tasks/cleanup $CONTAINER' EXIT
-if ! docker start $CONTAINER 2>/dev/null
-then
-	docker pull mongo >/dev/null
-        docker run -dp 27017:27017 --name $CONTAINER mongo >/dev/null
-	echo DB container $CONTAINER
-fi
+
 echo "Loading fixtures..."
 node tasks/fixtures.js >/dev/null
 echo "Running server..."
